requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
dateparser==1.2.0
python-dotenv==1.0.1
groq==0.33.0
loguru==0.7.2
